+++ src/components/create-post.jsx
@@ -18,7 +18,6 @@ import { ButtonLink } from './button-link';
 
 const attachmentsMaxCount = CONFIG.attachments.maxCount;
 
-const isTextEmpty = (text) => text == '' || /^\s+$/.test(text);
 const getDefaultState = (invitation = '') => ({
   isFormEmpty: true,
   isMoreOpen: false,
@@ -122,7 +121,7 @@ export default class CreatePost extends Component {
 
   checkCreatePostAvailability = () => {
     const isFormEmpty =
-      isTextEmpty(this.state.postText) || !this.selectFeeds || this.selectFeeds.values === 0;
+      this.state.postText.trim() === '' || !this.selectFeeds || this.selectFeeds.values === 0;
 
     this.setState({ isFormEmpty });
   };
@@ -172,7 +171,7 @@ export default class CreatePost extends Component {
 
   canSubmitForm = () => {
     return (
-      !this.state.isFormEmpty &&
+      (!this.state.isFormEmpty || this.state.attachments.length > 0) &&
       !this.state.attLoading &&
       !this.props.createPostViewState.isPending &&
       this.selectFeeds &&