+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.101.0] - Not released
+### Changed
+- Post can now have an empty body if it contains one or more attachments.
 
 ## [1.100.0] - 2021-07-30
 ### Added